CNDB-10221: use int for SAI postings (#1199)
After careful examination of the SAI code base, it is clear that we have a restriction that a single segment can have at most `Integer.MAX_VALUE` rows, and therefore, we can safely use `int` instead of `long` for postings. This PR proposes changes to make the code consistent, which should improve readability and future design, assuming that we want to continue to rely on having `Integer.MAX_VALUE` as the upper bound on number of rows in a segment.

The PR removes the `OffsetPostingList` because it incorrectly applies the segment offset in the `PostingList`, which should not happen.

Note that I was able to make all of the `long` to `int` conversions with only a handful of actual type conversions. I used `Math.toIntExact` to ensure we don't introduce any regressions. Additionally, I documented special cases that I thought were notable.

In the second commit, I updated part of the kd tree logic to account for the fact that doc ids (row ids) are always integers. We still use the `PackedLongValues` object though. Perhaps there is another data structure we could use?

+++ src/java/org/apache/cassandra/index/sai/disk/IndexSearcherContext.java
@@ -63,7 +63,7 @@ public class IndexSearcherContext
         return segmentRowIdOffset;
     }
 
-    long count()
+    int count()
     {
         return postingList.size();
     }
